Add files via upload
Project1.py ---> This Python code is a simple dice game that 2-4 players can play. Players take turns rolling dice to try to reach 50 points.

Project2.py ---> This Python code creates a story by taking specific words from the user. The words between the < > signs are used as placeholders and the user is asked to complete them.

Project3.py ---> This Python code creates a game that asks the user random math problems and checks their solutions. It is a simple quiz application that evaluates the user's performance by calculating the total time.